==> 1_collect_clean_AGE.ipynb <==
import pandas as pd
import os
import requests, zipfile, io
import numpy as np


# Age

age = pd.read_csv("data/ipo_age.csv")
age

age.columns = age.columns.str.replace(' ','_').str.lower()

age.drop(age.columns[np.r_[0,1,5,9]], axis=1, inplace=True)

age.info()

age = age[age['ticker'].notna()]
age.info()

**Manual Data Formatting**

age[age['ticker']=='BIGC'].replace('-99','2009',inplace=True)

age[age['ticker']=='ANNX'].replace('-99','2011',inplace=True)

age.replace(['9','.'], '0',inplace=True)


messy_columns = age[(age["vc_dummy"].isin(['9','.']))
   | (age["dual"].isin(['9','.']))
   | (age["rollup"].isin(['.']))
   | (age["internet"].isin(['9','.']))]

messy_columns[messy_columns["rollup"].isin(['9','.'])] 

<br>**vc_dummy** - is a dummy with 1 for VC and 2 for a subset of VC (growth capital)
<br>**dual** - dummy for multiple share class IPOs
<br>**rollup** - 0-1 dummy variable for whether the IPO was a rollup 
<br>**internet** - dummy with 1 if it is an internet-based  company 

age.info()

age.drop_duplicates(inplace=True)

age= age.groupby(['ticker']).max().reset_index()

age 

# Save Collected Data 

import joblib 
joblib.dump(age,"age.pkl")

ls
