==> 2_collect_clean_MANAGERS.ipynb <==
import pandas as pd
import os
import requests, zipfile, io
import numpy as np


# Managers

managers = pd.read_csv("data/ipo_managers.csv", header=[0,1])


managers.columns = ['_'.join(col) for col in managers.columns.values]


managers.columns = managers.columns.str.replace(' ','_').str.lower()

managers 

managers.rename(columns={
    'unnamed:_1_level_0_issuer':'company_name', 
    'unnamed:_2_level_0_symbol': 'ticker',
    'lead/joint-lead__managers': 'all_managers',
    'opening_price': 'open',
    '1st_day_close': 'close',
    'offer_price': 'offer',

    'unnamed:_11_level_0_performed': 'performed'},
                inplace=True)

managers.drop(managers.columns[np.r_[7:12]],inplace=True,axis=1)

managers.drop_duplicates(inplace=True)

managers= managers.groupby(['ticker']).max().reset_index()

**Data Formatting**

def format_prices(df,col):
    df[col] = df[col].str.strip('$\n\t').str.replace(',','').astype(float)
    return df[col]

cols = ['close', 'open','offer']

for col in cols:
    format_prices(managers,col)


managers['trade_date'] = pd.to_datetime(managers['trade_date'], errors = 'coerce')


# turn multi-valued managers cells into multi-row, single-valued --> -->
managers_columns = list(managers.columns)
managers_columns.remove('all_managers')
managers_columns

managers.dropna(subset=['all_managers'], inplace=True)
managers['all_managers'] = managers['all_managers'].map(lambda x: x.replace("/"," ,"))

managers['all_managers'] = managers['all_managers'].map(lambda x: x.split(","))
managers['all_managers']

managers_melt = managers.all_managers.apply(pd.Series) \
    .merge(managers, right_index = True, left_index = True) \
    .drop(["all_managers"], axis = 1) \
    .melt(id_vars = managers_columns , value_name = "managers") 


managers_melt.dropna(subset=['managers'], inplace=True)

managers_melt.drop(managers_melt.columns[np.r_[6]],inplace=True,axis=1)

managers_melt.info()

managers.info()

# Save Collected Data 

import joblib 
joblib.dump(managers_melt,"managers_melt.pkl")
joblib.dump(managers,"managers_all.pkl")


ls
