==> 3a_collect_SEC.ipynb <==
import pandas as pd
import os
import requests, zipfile, io
import numpy as np


# Financial Data

# import financials -- all num.txt and sub.txt
years = list(range(2009,2021))
quarters = ['q1','q2','q3','q4']
filenames = ['num.txt','sub.txt']   
dataset = 'financial-statement-data-sets'
# 'pre.txt','tag.txt

def SEC_data(dataset,quarter,year,filename):
    url = f"https://www.sec.gov/files/dera/data/{dataset}/{year}{quarter}.zip"
    r = requests.get(url)
    z = zipfile.ZipFile(io.BytesIO(r.content))
    z.extractall()
    df = pd.read_csv(filename,sep="\t",dtype={"cik":str})
    return df

financials = []    
for year in years:
    for quarter in quarters:
        for filename in filenames:
            df = SEC_data(dataset,quarter, year, filename)
            financials.append(df)

numi = []
subi = []
financials_all = []

for n in list(range(len(financials))):
    if n % 2 == 0:
        numi.append(n)
    else:  
        subi.append(n)
        
for n, s in zip(numi,subi):
    merged_df = financials[n].merge(financials[s][["adsh","cik","name","period","fy","fp","filed", "form","instance"]])
    financials_all.append(merged_df)

Filter for only 10-Ks only
<br> Companies have different FYs and larger companies take much longer to submit filings -- will not filter for quarter 

financials = pd.concat(financials_all)
financials.info()

tags = pd.DataFrame(financials['tag'].value_counts())
tags[tags['tag']>=500000]

financials_tags = financials[financials["tag"].isin(["NetIncomeLoss",
                                                 "CashAndCashEquivalentsAtCarryingValue",
                                                 "OperatingIncomeLoss",
                                                 "StockholdersEquity",
                                                 "Assets",
                                                 
                                                 "LiabilitiesAndStockholdersEquity",
                                                 "LiabilitiesAndStockholdersEquity",
                                                 
                                                 "EarningsPerShareBasic",
                                                 "EarningsPerShareDiluted",
                                                 
                                                 "NetCashProvidedByUsedInOperatingActivities",
                                                 "NetCashProvidedByUsedInFinancingActivities",
                                                 
                                                 "RetainedEarningsAccumulatedDeficit",
                                                 "EarningsPerShareDiluted",
                                                 "CommonStockSharesAuthorized"
                                                 ])]

annual_raw = financials_tags[financials_tags["form"].isin(["10-K","20-K","40-K"])]


annual_raw.drop(annual_raw.columns[np.r_[7]], axis=1, inplace=True)


annual_raw.drop_duplicates(inplace= True)

annual_raw.info()

annual = annual_raw.pivot_table(values='value', 
                          index= ['version',
                                  'coreg',
                                  'ddate',
                                  'qtrs',
                                  'uom',
                                  'adsh',
                                  'cik',
                                  'name',
                                   'filed',
                                  'form',
                                  'instance'],
                          columns = ['tag','fy'])


annual.reset_index(inplace=True)

annual

pd.set_option('display.max_columns', None)


annual.describe()

# Crowdfunding Data

# import crowdfunding -- 'FORM_C_SUBMISSION.tsv','FORM_C_DISCLOSURE.tsv'
quarters = ['q1_cf','q2_cf','q3_cf','q4_cf']
years = list(range(2017,2021))
dataset = 'crowdfunding-offerings-data-sets'

crowdfunding = []    
for year in years:
    for quarter in quarters:
        file_disc = f"{year}{quarter.capitalize()}/FORM_C_DISCLOSURE.tsv"
        file_issuer = f"{year}{quarter.capitalize()}/FORM_C_ISSUER_INFORMATION.tsv"        
        filenames =[file_disc,file_issuer]
        for filename in filenames:
            df = SEC_data(dataset, quarter, year, filename)
            crowdfunding.append(df)
 

disc_i = []
issuer_i = []
crowdfunding_all = []

for n in list(range(len(crowdfunding))):
    if n % 2 == 0:
        disc_i.append(n)
    else:  
        issuer_i.append(n)
        
for n, s in zip(disc_i,issuer_i):
    df = crowdfunding[n].merge(crowdfunding[s])
    crowdfunding_all.append(df)

crowdfunding = pd.concat(crowdfunding_all)

crowdfunding.info()

crowdfunding.drop(crowdfunding.columns[np.r_[30:45,46,48,49]], axis=1, inplace=True)

crowdfunding.info()

crowdfunding.rename(columns={
    'ACCESSION_NUMBER':'accession_number',
    'COMPENSATIONAMOUNT': 'compensation_amount',
    'FINANCIALINTEREST': 'financial_interest',
    'SECURITYOFFEREDTYPE': 'security_offered_type',
    'SECURITYOFFEREDOTHERDESC': 'security_offered_other_desc',
    'NOOFSECURITYOFFERED':'no_of_security_offered', 
    'PRICE': 'price',
    'PRICEDETERMINATIONMETHOD': 'price_determination_method',
    'OFFERINGAMOUNT': 'offering_amount',
    'DESCOVERSUBSCRIPTION' : 'desc_over_subscription',
   
    'OVERSUBSCRIPTIONACCEPTED': 'oversubscription_accepted',    
    'OVERSUBSCRIPTIONALLOCATIONTYPE':'oversubscriptional_location_type', 
    'COMPENSATIONAMOUNT': 'des_cover_subscription',
    'MAXIMUMOFFERINGAMOUNT': 'maximum_offering_amount',
    'DEADLINEDATE': 'deadline_date',
    
    'CURRENTEMPLOYEES': 'current_employees',
    
    
    'TOTALASSETMOSTRECENTFISCALYEAR':'total_asset_most_recent_fiscalyear', 
    'TOTALASSETPRIORFISCALYEAR': 'total_asset_prior_fiscal_year',
    
    'CASHEQUIMOSTRECENTFISCALYEAR': 'cash_equi_most_recent_fiscal_year',
    'CASHEQUIPRIORFISCALYEAR': 'cash_equi_prior_fiscal_year',
    
    'ACTRECEIVEDRECENTFISCALYEAR': 'act_received_recent_fiscal_year',
    'ACTRECEIVEDPRIORFISCALYEAR':'act_received_prior_fiscal_year', 
    
    'SHORTTERMDEBTMRECENTFISCALYEAR': 'short_term_debt_m_recent_fiscal_year',
    'SHORTTERMDEBTPRIORFISCALYEAR': 'short_term_debt_prior_fiscal_year',
    
    'LONGTERMDEBTRECENTFISCALYEAR': 'long_term_debt_recent_fiscal_year',
    'LONGTERMDEBTPRIORFISCALYEAR': 'long_term_debt_prior_fiscal_year',    
    
    'REVENUEMOSTRECENTFISCALYEAR':'revenue_most_recent_fiscal_year', 
    'REVENUEPRIORFISCALYEAR': 'revenue_prior_fiscal_year',
    
    'COSTGOODSSOLDRECENTFISCALYEAR': 'cost_goods_sold_recent_fiscal_year',
    'COSTGOODSSOLDPRIORFISCALYEAR': 'cost_goods_sold_prior_fiscal_year',
    
    'TAXPAIDMOSTRECENTFISCALYEAR': 'tax_paid_most_recent_fiscal_year',
    
    'ZIPCODE':'zipcode', 
    'COMPANYNAME': 'company_name',
    'CRDNUMBER': 'financial_interest'
},
                inplace=True)

crowdfunding.columns

# Save Collected Data 

import joblib 
joblib.dump(crowdfunding,"crowdfunding.pkl")
joblib.dump(annual,"sec_10k_raw.pkl")

ls
