==> 3b_clean_SEC.ipynb <==
import pandas as pd
import joblib 
import numpy as np 
import datetime as datetime
import seaborn as sns
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline 
import re
from datetime import datetime


**Only cleaning `SEC 10K` data since `crowdfunding` data proved to not be useful for this project**

# crowdfunding = joblib.load("crowdfunding.pkl")
annual = joblib.load("sec_10k_raw.pkl")

unique_co = pd.DataFrame(annual['name'].value_counts()).reset_index().sort_values(by ='name',ascending=False)
unique_co

def extract_ticker(df,col):
    df[col] = df[col].map(lambda x: x.split('-')[0])
    df[col] = df[col].map(lambda x: re.findall('\d*\D+',x)[0])
    df[col] = df[col].map(lambda x: x.upper())
        
    return df[col]
    

annual['instance'] = extract_ticker(annual,'instance')

annual['fp'].value_counts()

annual.rename(columns={'instance':'ticker',
'ddate':'end_date',
'period':'balance_sheet_date',
'fy':'fiscal_year',
'filed':'filed_date',
'Assets':'assets',
'CashAndCashEquivalentsAtCarryingValue':'cash_and_cash_equivalent_sat_carrying_value',
'CommonStockSharesAuthorized':'common_stock_shares_authorized',
'EarningsPerShareBasic':'earnings_per_share_basic',
'EarningsPerShareDiluted':'earnings_per_share_diluted',
'LiabilitiesAndStockholdersEquity':'liabilities_and_stock_holders_equity',
'NetCashProvidedByUsedInFinancingActivities':'net_cash_provided_by_used_in_financing_activities',
'NetCashProvidedByUsedInOperatingActivities':'net_cash_provided_by_used_in_operating_activities',
'NetIncomeLoss':'net_income_loss',
'OperatingIncomeLoss':'operating_incomeloss',
'RetainedEarningsAccumulatedDeficit':'retained_earnings_accumulated_deficit',
'StockholdersEquity':'stockholders_equity'},inplace=True)

pd.set_option('display.max_columns',500)
annual

**Data Formatting**

def date_format(df,col):
    df[col] = df[col].astype(str) 
    df[col]= df[col].map(lambda x: datetime.strptime(x, '%Y%m%d').strftime('%Y-%m-%d'))
    df[col] = pd.to_datetime(df[col], errors = 'coerce')
    return df[col]

def year_format(df,col):
    df[col] = df[col].astype(int)
    return df[col]
 

date_format(annual,'end_date')

date_format(annual,'filed_date')

year_format(annual,'fiscal_year') 

annual.info()

annual_test = annual.drop(annual.columns[np.r_[0,1,3,4,5,6,7,8,10,12,16:19,24]], axis=1)
annual_test                  

joblib.dump(annual_test,"sec_10k.pkl")
