==> 4a_collect_YF1.ipynb <==
import joblib
import pandas as pd
import datetime
import yfinance as yf


tickers = []

cols = ['symbol',
'sector',        
'totalRevenue',
'fullTimeEmployees',
'ebitdaMargins',
'profitMargins',
'grossMargins',
'operatingCashflow',
'revenueGrowth',
'operatingMargins',
'ebitda',
'grossProfits',
'bookValue',
'nextFiscalYearEnd',
'enterpriseToRevenue',
'enterpriseValue',
'marketCap']
df = pd.DataFrame(columns=cols)


 
symbols = []    
for ticker in tickers:
    ticker = yf.Ticker(ticker)
    symbols.append(ticker)
symbols    

        
rows = []
rows_processed=0


for symbol in symbols:
#     clear_output(wait=True)
    values = []
    try:
        
        for col in cols:
            if col not in list(symbol.info.keys()):
                for val in list(symbol.info.values()):
                    val = 'Null'

            else:
                for val in list(symbol.info.values()):
                    val = symbol.info[col]
            values.append(val)

        rows.append(values)
        rows_processed += 1

    except  (RuntimeError, IndexError, TypeError,KeyError):
        pass
    print('Progress: {}/{} rows processed'.format(rows_processed, len(tickers)) )
    print(f"""current progress : {rows_processed/len(tickers)*100} %""" )

    print(f"""finished {symbol} at {str(datetime.datetime.now())}""" )    
    
df = df.append(pd.DataFrame(rows, columns=cols))
df


df.to_csv('data/yahoo_finance/ ')
