==> 4b_clean_YF1.ipynb <==
import pandas as pd 
import numpy as np 
import joblib

**First Phase Data**

numbers = ['0500','1000','1500','2000','2500','3000','3500','4000','4500','5000','5500']
files = []
for num in numbers: 
    file = pd.read_csv(f'data/yahoo_finance/financials_raw{num}')
    files.append(file)


financials_raw = pd.concat(files)
financials_raw.info()

financials_raw.drop(financials_raw.columns[np.r_[0,2,15,16,18,19,20]],inplace=True, axis=1)

financials_raw.info()

**Second Phase Data**

numbers = ['0500','1000','1500','2000','2500','3000','3500','4000','4500','5000',
           '5500','5500','6000','6500','7000','7500','8000','9000','9500', '10000','10500','11000']

files = []
for num in numbers: 
    file = pd.read_csv(f'data/yahoo_finance/yf1_{num}')
    files.append(file)


yf_raw = pd.concat(files)
yf_raw.info()
 



yf_raw.drop(yf_raw.columns[np.r_[0]],inplace=True, axis=1)

yf_raw.info()

**Combine first and second phase data**

df12 = [yf_raw,financials_raw]
yf1 = pd.concat(df12)
yf1.info()

yf1.rename(columns={
    'symbol':'ticker',
    'totalRevenue':'total_revenue',
    'fullTimeEmployees': 'full_time_employees',
    'ebitdaMargins': 'ebitda_margins',
    'profitMargins':'profit_margins',
    'grossMargins': 'gross_margins',
    'operatingCashflow': 'operating_cashflow',
    'revenueGrowth':'revenue_growth',
    'operatingMargins': 'operating_margins',
    'grossProfits': 'gross_profits',
    'bookValue': 'book_value',    
    'netIncomeToCommon': 'net_income_to_common',
    'enterpriseToRevenue': 'enterprise_to_revenue',
    'enterpriseValue': 'enterprise_value',
    'marketCap': 'market_cap'
},
                inplace=True)


yf1 = yf1[yf1['ticker']!='Null']


def to_float(df,col):        
    df[col] = df[col].replace('Null',0).astype(float)
    return df[col]

def to_int(df,col):       
    df[col] = df[col].fillna(0).replace('Null',0).astype(int)
    return df[col]

floats=['ebitda_margins',
        'profit_margins',
        'gross_margins',
        'operating_margins',
        'revenue_growth',
        'book_value',
        'enterprise_to_revenue',
       'price_to_book']

ints=['total_revenue',
      'operating_cashflow',
      'full_time_employees',
      'ebitda',
      'gross_profits',
      'net_income_to_common',
      'enterprise_value',
      'market_cap',
     'shares_short']

for col in list(yf1.columns):
    if col in floats: 
        to_float(yf1,col)
        
    elif col in ints: 
        to_int(yf1,col)
    else:
        None
        


yf1.drop(columns='nextFiscalYearEnd',inplace=True,axis=1)

**Handling duplicates**

yf1.drop_duplicates(inplace=True)

print(f"{len(set(list(yf1['ticker'].unique())))} unique tickers and {len(yf1)} rows in YF1 dataset  ")


yf1[yf1.duplicated(['ticker'], keep=False)]


yf1[yf1['ticker']=='DCT']

yf1= yf1.groupby(['ticker']).max().reset_index()

print(f"{len(set(list(yf1['ticker'].unique())))} unique tickers and {len(yf1)} rows in YF1 dataset  ")


**Test which columns to keep based on what has the most data**

f2 = yf1[yf1['profit_margins'] >0]
f2.info()

yf91 = yf1[(yf1['enterprise_to_revenue'] ==0)]
#                                &(financials_raw['ebitda_margins'] >0)
#                    &(yf1['enterprise_to_revenue'] >0)
         

yf91.describe() 

yf1= yf1[['total_revenue','operating_cashflow','full_time_employees','ticker','sector','revenue_growth','gross_margins','profit_margins','enterprise_to_revenue','ebitda_margins','market_cap']]

yf1['sector'] = yf1['sector'].replace('Null', 'Other')

joblib.dump(yf1,"yf1.pkl")
