==> 5a_collect_YF2.ipynb <==
import yfinance as yf
from yahoofinancials import YahooFinancials
import pandas as pd
import numpy as np
import joblib 
from pathlib import Path

import datetime


tickers =[]
tickers


cols = ['date',
'high',
'low',        
'open',
'close',
'volume',
'formatted_date']  
    
df = pd.DataFrame(columns=cols)

 

from interruptingcow import timeout 


min_data_list = []

for ticker in tickers:
    try: 
        with timeout(7, exception = RuntimeError):  
            symbol = YahooFinancials(ticker)
            
            data = symbol.get_historical_price_data("1850-01-01", "2021-05-30", "monthly")

            min_data = data[ticker]['prices'][6]
            min_data['ticker'] = ticker
            min_data_list.append(min_data)
            print(f"""finished {ticker} at {str(datetime.datetime.now())}""" ) 
    
    except  (RuntimeError, IndexError, TypeError,KeyError):
        pass


df1 = pd.DataFrame(min_data_list)
df1
 

df1.to_csv('data/yahoo_ipo/ ')
