==> 6_combine_2.ipynb <==
import pandas as pd
import joblib 
import numpy as np 
import datetime as datetime
import seaborn as sns
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline  
import joblib
from pathlib import Path



file_path = Path("/Users/lorelablaka/Documents/Flatiron/phase-5/capstone/collect_clean")

age = joblib.load(f"{file_path}/age.pkl")
sec_10k = joblib.load(f"{file_path}/sec_10k.pkl")
managers = joblib.load(f"{file_path}/managers_all.pkl")
yf1 = joblib.load(f"{file_path}/yf1.pkl")
yf2 = joblib.load(f"{file_path}/yf2.pkl")

**EDA of Datasets**

dfs = [age,sec_10k,managers,yf1,yf2]
df_names = ['age','sec_10k','managers','yf1','yf2']

for df in dfs:
    unique_list = list(df['ticker'].unique())
    print(f"{len(set(unique_list))}  unique tickers and {len(df)} rows in dataset")
    
    
 


**Concatinate `yahoofinance` IPO data with `managers` IPO data**

df_man = managers[['ticker','trade_date','open','close']]
df_man.info()

def ipo_returns(df,col1,col2):
    ret = (df[col2] - df[col1] )/df[col1]
    return ret

df_man['day_1_%'] = ipo_returns(df_man,'open','close')

df_man =df_man[['ticker','trade_date','day_1_%']]

all_ipos = (yf2,df_man)
all_ipos = pd.concat(all_ipos)
all_ipos.info()

Significantly less rows for `month_7_%`  

all_ipos= all_ipos.groupby(['ticker']).max().reset_index()
all_ipos

**Merging `all_ipo` data with recent company financials from `yf1` dataset**

yf =  all_ipos.merge(yf1, how='left',on='ticker')
yf.info()

yf['ipo_year']= yf['trade_date'].dt.year
yf['ipo_year']

**Merging `all_ipo` data and recent company financials with `age` dataset to include stock features**

data2 = yf.merge(age, how='left',on=['ticker'])
data2.info()

Clean up `data2` and create `years_to_ipo` feature

data2['ipo_year'] = data2['ipo_year'].fillna(2000).astype(int)


data2['founding'] = data2['founding'].fillna(2000).astype(int)


data2['years_to_ipo'] = data2['ipo_year'] - data2['founding']


**Modeling 2 Dataset**

data2 = data2[['ticker',
               'trade_date', 'ipo_year',
                      'day_1_%',
                      'month_7_%',
                      'sector',
                      'vc_dummy',
                      'rollup',
                      'dual',
                      'internet',
                      'years_to_ipo',
                      'enterprise_to_revenue',
                      'ebitda_margins',
                      'profit_margins',
                      'gross_margins',
                      'revenue_growth',
                      'total_revenue',
                      'operating_cashflow'
                    ]]
data2.info()


data2= data2.groupby(['ticker']).max().reset_index()

data2 = data2[data2['vc_dummy'].notna()]


data2 = data2[data2['day_1_%'].replace([np.inf, -np.inf], np.nan).notna()]
data2.info()

**Feature Engineer Dependent Variable `Successful`**

data2['gross_margins'].describe().T

data2['enterprise_to_revenue'].describe().T

def profitable(value):
    if value >=.65:
        return 1
    else:
        return 0
data2['profitable'] = data2['gross_margins'].map(lambda x: profitable(x))


data2['profitable'].value_counts()  

def valuable(value):
    if value >=4.27:
        return 1
    else:
        return 0
data2['valuable'] = data2['enterprise_to_revenue'].map(lambda x: valuable(x))


data2['valuable'].value_counts()

data2['day_1_%'] = data2['day_1_%'].replace([np.inf, -np.inf],0.0)

data2['month_7_%'] = data2['month_7_%'].replace([np.inf, -np.inf],0.0)

data2['success']  = data2['valuable']+data2['profitable']
data2['success'].value_counts()

def success(value):
    if value > 0:
        return 1
        
    if value ==0:
        return 0

    
data2['successful'] = data2['success'].map(lambda x: success(x))

data2['successful'].value_counts()

data2 = data2[data2['successful'].notna()]


data2 = data2[data2['sector'].notna()]


data2 = data2[data2['enterprise_to_revenue'].notna()]


data2.info()

data2_viz = data2

data2 = data2[['ticker',
                      'sector',
                      'vc_dummy',
                      'rollup',
                      'dual',
                      'internet',
                      'years_to_ipo',
                      'day_1_%',
                      'month_7_%',
                      'successful']]
data2.info()

joblib.dump(data2_viz,"data2_viz.pkl")

joblib.dump(data2,"data2.pkl")
